requirements.txt
asgiref==3.2.7
astroid==2.4.0
autopep8==1.5.2
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.5
django-searchable-encrypted-fields==0.1.4
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.9.7
pylint==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.12.1
